+++ src/extension.ts
@@ -15,6 +15,9 @@ import { ApolloStudioGraphsTreeDataProvider, StudioGraphTreeItem, StudioGraphVar
 import { getSelectedWorkbenchFile, saveWorkbenchFile, workspaceQueriesFolderPath, writeLocalSchemaToFile } from './helpers';
 import { deleteSchemaFile } from './workbench/current-workbench-schemas/deleteSchemaFile';
 import { ApolloStudioGraphOpsTreeDataProvider, StudioOperationTreeItem } from './workbench/studio-operations/apolloStudioGraphOpsTreeDataProvider';
+import { StateManager } from './workbench/stateManager';
+import { getGraphSchemasByVariant } from './studio-gql/graphClient';
+import { GetGraphSchemas_service_implementingServices_FederatedImplementingServices } from './studio-gql/types/GetGraphSchemas';
 
 export class ApolloWorkbench {
 	graphName: string = "";
@@ -38,37 +41,59 @@ export function deactivate(context: vscode.ExtensionContext): undefined {
 	return undefined;
 }
 
-export function activate(context: vscode.ExtensionContext) {
-	const localWorkbenchFilesProvider = new LocalWorkbenchFilesTreeDataProvider(vscode.workspace.rootPath ?? ".");
-	const currentWorkbenchSchemasProvider = new CurrentWorkbenchSchemasTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
-	const currentWorkbenchOperationsProvider = new CurrentWorkbenchOpsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
-	const apolloStudioGraphsProvider = new ApolloStudioGraphsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
-	const apolloStudioGraphOpsProvider = new ApolloStudioGraphOpsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+export async function activate(context: vscode.ExtensionContext) {
+	const ensureFolderIsOpenommand = vscode.commands.registerCommand('extension.ensureFolderIsOpen', async () => {
+		if (!vscode.workspace.rootPath) {
+			let openFolder = "Open Folder";
+			let response = await vscode.window.showErrorMessage("You must open a folder to use Apollo Workbench. We'll need a place to store a .workbench folder that holds schema/query files.", openFolder);
+			if (response == openFolder)
+				await vscode.commands.executeCommand('extension.openFolder');
+		} else {
+			setupApolloWorkbench(context);
+		}
+	});
+	const openCodeFolderCommand = vscode.commands.registerCommand('extension.openFolder', async () => {
+		let folder = await vscode.window.showOpenDialog({ canSelectFiles: false, canSelectFolders: true, canSelectMany: false });
+		if (folder)
+			await vscode.commands.executeCommand('vscode.openFolder', folder[0]);
+	});
+	context.subscriptions.push(openCodeFolderCommand);
+	context.subscriptions.push(ensureFolderIsOpenommand);
+	vscode.commands.executeCommand('extension.ensureFolderIsOpen');
+}
+
+function setupApolloWorkbench(context: vscode.ExtensionContext) {
+	// const localWorkbenchFilesProvider = new LocalWorkbenchFilesTreeDataProvider(vscode.workspace.rootPath ?? ".");
+	// const currentWorkbenchSchemasProvider = new CurrentWorkbenchSchemasTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+	// const currentWorkbenchOperationsProvider = new CurrentWorkbenchOpsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+	// const apolloStudioGraphsProvider = new ApolloStudioGraphsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+	// const apolloStudioGraphOpsProvider = new ApolloStudioGraphOpsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
 	context.subscriptions.push(compositionDiagnostics);
 
 	fileWatchManager.extensionContext = context;
-	fileWatchManager.localWorkbenchFilesProvider = localWorkbenchFilesProvider;
-	fileWatchManager.currentWorkbenchSchemasProvider = currentWorkbenchSchemasProvider;
-	fileWatchManager.currentWorkbenchOperationsProvider = currentWorkbenchOperationsProvider;
-	fileWatchManager.apolloStudioGraphsProvider = apolloStudioGraphsProvider;
-	fileWatchManager.apolloStudioGraphOpsProvider = apolloStudioGraphOpsProvider;
+	StateManager.context = context;
+	StateManager.localWorkbenchFilesProvider = new LocalWorkbenchFilesTreeDataProvider(vscode.workspace.rootPath ?? ".");
+	StateManager.currentWorkbenchSchemasProvider = new CurrentWorkbenchSchemasTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+	StateManager.currentWorkbenchOperationsProvider = new CurrentWorkbenchOpsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+	StateManager.apolloStudioGraphsProvider = new ApolloStudioGraphsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
+	StateManager.apolloStudioGraphOpsProvider = new ApolloStudioGraphOpsTreeDataProvider(vscode.workspace.rootPath ?? ".", context);
 
 	//Global Extension Commands
-	const newWorkbenchCommand = vscode.commands.registerCommand('extension.newWorkbench', async () => { newWorkbench(localWorkbenchFilesProvider) });
+	const newWorkbenchCommand = vscode.commands.registerCommand('extension.newWorkbench', async () => { newWorkbench(StateManager.localWorkbenchFilesProvider) });
 	const startWorkbenchCommand = vscode.commands.registerCommand('extension.startWorkbench', async () => await fileWatchManager.start());
 	const stopWorkbenchCommand = vscode.commands.registerCommand('extension.stopWorkbench', async () => await fileWatchManager.stop())
 	const enterStudioApiKeyCommand = vscode.commands.registerCommand('extension.enterStudioApiKey', async () => {
 		let apiKey = await vscode.window.showInputBox({ placeHolder: "Enter User API Key - user:gh.michael-watson:023jr324tj...." })
 		if (apiKey) {
 			context.globalState.update('APOLLO_KEY', apiKey);
-			apolloStudioGraphsProvider.refresh();
+			StateManager.apolloStudioGraphsProvider.refresh();
 		}
 	})
 	vscode.commands.registerCommand('extension.deleteStudioApiKey', () => {
 		context.globalState.update('APOLLO_KEY', "");
 		context.globalState.update('APOLLO_SELCTED_ACCOUNT', "");
-		apolloStudioGraphsProvider.refresh();
-		apolloStudioGraphOpsProvider.refresh();
+		StateManager.apolloStudioGraphsProvider.refresh();
+		StateManager.apolloStudioGraphOpsProvider.refresh();
 	});
 
 	context.subscriptions.push(newWorkbenchCommand);
@@ -77,19 +102,19 @@ export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(enterStudioApiKeyCommand);
 
 	//Current Loaded Workbench Schemas Commands
-	vscode.window.registerTreeDataProvider('current-workbench-schemas', currentWorkbenchSchemasProvider);
-	vscode.commands.registerCommand('current-workbench-schemas.addSchema', async () => await createSchemaFile(context, currentWorkbenchSchemasProvider));
+	vscode.window.registerTreeDataProvider('current-workbench-schemas', StateManager.currentWorkbenchSchemasProvider);
+	vscode.commands.registerCommand('current-workbench-schemas.addSchema', async () => await createSchemaFile(context, StateManager.currentWorkbenchSchemasProvider));
 	vscode.commands.registerCommand("current-workbench-schemas.editSchema", editSchema);
-	vscode.commands.registerCommand("current-workbench-schemas.deleteSchema", async (serviceToDelete: WorkbenchSchemaTreeItem) => await deleteSchemaFile(serviceToDelete.serviceName, context, currentWorkbenchSchemasProvider));
-	vscode.commands.registerCommand('current-workbench-schemas.refreshSchemas', async () => currentWorkbenchSchemasProvider.refresh());
+	vscode.commands.registerCommand("current-workbench-schemas.deleteSchema", async (serviceToDelete: WorkbenchSchemaTreeItem) => await deleteSchemaFile(serviceToDelete.serviceName, context, StateManager.currentWorkbenchSchemasProvider));
+	vscode.commands.registerCommand('current-workbench-schemas.refreshSchemas', async () => StateManager.currentWorkbenchSchemasProvider.refresh());
 
 
 	//Current Loaded Workbench Operations Commands
-	vscode.window.registerTreeDataProvider('current-workbench-operations', currentWorkbenchOperationsProvider);
+	vscode.window.registerTreeDataProvider('current-workbench-operations', StateManager.currentWorkbenchOperationsProvider);
 	vscode.commands.registerCommand('current-workbench-operations.addOperation', async () => await fileWatchManager.createOperation(context));
 	vscode.commands.registerCommand("current-workbench-operations.editOperation", async (operation: WorkbenchOperationTreeItem) => await fileWatchManager.editOperation(operation.operationName));
 	vscode.commands.registerCommand("current-workbench-operations.deleteOperation", async (operation: WorkbenchOperationTreeItem) => fileWatchManager.deleteOperation(operation.operationName));
-	vscode.commands.registerCommand('current-workbench-operations.refreshOperations', async () => currentWorkbenchOperationsProvider.refresh());
+	vscode.commands.registerCommand('current-workbench-operations.refreshOperations', async () => StateManager.currentWorkbenchOperationsProvider.refresh());
 	vscode.commands.registerCommand('current-workbench-operations.openQueryPlan', async (op: StudioOperationTreeItem) => {
 		outputChannel.appendLine(`Opening query plan for operation ${op.operationName}`);
 		const workbenchQueriesFolder = workspaceQueriesFolderPath();
@@ -98,20 +123,20 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 
 	//Local Workbench Files Commands
-	vscode.window.registerTreeDataProvider('local-workbench-files', localWorkbenchFilesProvider);
+	vscode.window.registerTreeDataProvider('local-workbench-files', StateManager.localWorkbenchFilesProvider);
 	vscode.commands.registerCommand("local-workbench-files.loadFile", (item: WorkbenchFileTreeItem) => {
 		outputChannel.appendLine(`Loading WB:${item.graphVariant} - ${item.filePath}`);
 		context.workspaceState.update("selectedWbFile", { name: item.graphVariant, path: item.filePath } as WorkbenchFile);
-		currentWorkbenchSchemasProvider.refresh();
-		currentWorkbenchOperationsProvider.refresh();
+		StateManager.currentWorkbenchSchemasProvider.refresh();
+		StateManager.currentWorkbenchOperationsProvider.refresh();
 	});
 	vscode.commands.registerCommand("local-workbench-files.deleteFile", async (item: WorkbenchFileTreeItem) => await fileWatchManager.deleteWorkbenchFile(item.filePath));
-	vscode.commands.registerCommand('local-workbench-files.refresh', async () => localWorkbenchFilesProvider.refresh());
+	vscode.commands.registerCommand('local-workbench-files.refresh', async () => StateManager.localWorkbenchFilesProvider.refresh());
 
 
 	//Apollo Studio Graphs Commands
-	vscode.window.registerTreeDataProvider('studio-graphs', apolloStudioGraphsProvider);
-	vscode.commands.registerCommand('studio-graphs.refresh', () => apolloStudioGraphsProvider.refresh());
+	vscode.window.registerTreeDataProvider('studio-graphs', StateManager.apolloStudioGraphsProvider);
+	vscode.commands.registerCommand('studio-graphs.refresh', () => StateManager.apolloStudioGraphsProvider.refresh());
 	vscode.commands.registerCommand('studio-graphs.createWorkbenchFromGraph', async (graphTreeItem: StudioGraphTreeItem) => {
 		let graphVariants: string[] = new Array<string>();
 		let graphVariantTreeItems: { [key: string]: StudioGraphVariantTreeItem } = {};
@@ -141,7 +166,7 @@ export function activate(context: vscode.ExtensionContext) {
 			});
 
 			saveWorkbenchFile(workbenchFile);
-			localWorkbenchFilesProvider.refresh();
+			StateManager.localWorkbenchFilesProvider.refresh();
 		}
 	});
 	vscode.commands.registerCommand('studio-graphs.createWorkbenchFromGraphWithVariant', async (graphVariantTreeItem: StudioGraphVariantTreeItem) => {
@@ -161,7 +186,7 @@ export function activate(context: vscode.ExtensionContext) {
 		});
 
 		saveWorkbenchFile(workbenchFile);
-		localWorkbenchFilesProvider.refresh();
+		StateManager.localWorkbenchFilesProvider.refresh();
 	});
 	vscode.commands.registerCommand('studio-graphs.addServiceSchemaToCurrentWorkbench', async (graphVariantServiceTreeItem: StudioGraphVariantServiceTreeItem) => {
 		let workbenchFile = getSelectedWorkbenchFile(context);
@@ -169,18 +194,45 @@ export function activate(context: vscode.ExtensionContext) {
 			writeLocalSchemaToFile(graphVariantServiceTreeItem.name, graphVariantServiceTreeItem.sdl);
 			workbenchFile.schemas[graphVariantServiceTreeItem.name] = graphVariantServiceTreeItem.sdl;
 			saveWorkbenchFile(workbenchFile);
-			currentWorkbenchSchemasProvider.refresh();
+			StateManager.currentWorkbenchSchemasProvider.refresh();
 		} else
 			vscode.window.showErrorMessage("There is no workbench file currently selected to add this schema to. Please select a local workbench file and then try again,");
 	});
 	vscode.commands.registerCommand('studio-graphs.loadOperations', async (graphTreeItem: StudioGraphTreeItem) => {
 		context.globalState.update("APOLLO_SELCTED_GRAPH_ID", graphTreeItem.graphId);
-		apolloStudioGraphOpsProvider.refresh();
+		StateManager.apolloStudioGraphOpsProvider.refresh();
 	});
+	vscode.commands.registerCommand('studio-graphs.switchOrg', async () => StateManager.setAccountId());
 
 	//Apollo Studio Graph Operations Commands
-	vscode.window.registerTreeDataProvider('studio-operations', apolloStudioGraphOpsProvider);
+	vscode.window.registerTreeDataProvider('studio-operations', StateManager.apolloStudioGraphOpsProvider);
 	vscode.commands.registerCommand('studio-operations.addToWorkbench', async (op: StudioOperationTreeItem) => {
 		await fileWatchManager.createOperation(context, op.operationName, op.operationSignature)
 	});
+
+
+	// vscode.commands.registerCommand('test', async (item: StudioGraphVariantTreeItem) => {
+	// 	let apiKey = StateManager.context.globalState.get("APOLLO_KEY") as string;
+	// 	let graphVariantServiceTreeItems = new Array<StudioGraphVariantServiceTreeItem>();
+
+	// 	//Query Studio for graph by variant
+	// 	let variantServices = await getGraphSchemasByVariant(apiKey, item.graphId, item.graphVariant);
+	// 	let implementingServices = variantServices.service?.implementingServices as GetGraphSchemas_service_implementingServices_FederatedImplementingServices;
+
+	// 	if (implementingServices) {
+	// 		//Loop through implemnting services and add to return objects
+	// 		for (var l = 0; l < implementingServices.services.length; l++) {
+	// 			let implemntingService = implementingServices.services[l];
+	// 			graphVariantServiceTreeItems.push(new StudioGraphVariantServiceTreeItem(item.graphId, item.graphVariant, implemntingService.name, implemntingService.activePartialSchema.sdl));
+	// 		}
+	// 	}
+	// 	else {
+	// 		let schema = variantServices.service?.schema?.document;
+	// 		graphVariantServiceTreeItems.push(new StudioGraphVariantServiceTreeItem(item.graphId, item.graphVariant, 'monolith-schema', schema));
+	// 	}
+	// 	// Set the implementing service tree items on the return objects
+	// 	item.children = graphVariantServiceTreeItems;
+
+	// 	item.getChildren();
+	// });
 }